dpi: remote_bitbang_sv: Introduce place holder for logic signals
When it comes to actual compile, seems we might have issues of
assignment to registers as not being permitted, so introduce a place
holder such as tms_o that can be assigned. The registers such as
jtag_tms_o is prevented from being assigned in multiplace places in the
module.

Signed-off-by: Kregg Castillo <kregg.castillo@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>