dpi: remote_jtag: Enable/Disable the server based on signal
Since we have an enable signal, we should enable/disable server socket
based on the enable signal. if the enable is asserted, we can enable the
server, else we should be disabling the server, if it existed.

We dont really want our server to be polled prior to the server being
ready either, so maintain a local state machine to track that
information as well.

Signed-off-by: Nishanth Menon <nm@ti.com>